+++ HISTORY.rst
@@ -3,6 +3,14 @@
 History
 -------
 
+1.1.0 (2016-04-20)
+++++++++++++++++++
+
+* Introduced HTML Formatter
+* Added log text escaping (closed #3)
+* Added requests timeout setting (closed  #1)
+* Catching and logging all requests and their exceptions (closed #2)
+
 1.0.0 (2016-04-19)
 ++++++++++++++++++